% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircaN_library.R
\name{circan}
\alias{circan}
\title{circan}
\usage{
circan(data, s2c, shiny = FALSE, mode = "default", init_value = 24,
  max_per = Inf, min_per = -Inf)
}
\arguments{
\item{data}{Dataframe containing the expression data. Samples must be in columns and genes in rows.
For an example see data(expression_example).}

\item{s2c}{Dataframe containing the metadata for the samples. Must have at least a 'sample' column
with the sample name as it appears in the data matrix;
a 'time' column with the time point the sample was collected;
and an 'ind' column containing information for the individual the sample comes from.
For an example see data(metadata_example).}

\item{shiny}{Is the package running in a shiny app? default to FALSE.}

\item{mode}{Algorithm to use in the NLS regression. Must be one of 'default' for Gauss-Newton, 'plinear' for the Golub-Pereyra algorithm
for partially linear least-squares models and 'port' for the ‘nl2sol’ algorithm from the Port library. Default is default. See nls documentation
for extended info.}

\item{init_value}{Initial value for the period. Default is set to 24.}

\item{max_per}{Maximum period to regress. Default is set to Inf.}

\item{min_per}{Minimum period to regress. Default is set to -Inf.}
}
\description{
Nonlinear least squares model for accurate detection of circadian expression patterns.
}
\examples{
# This runs CircaN on the example data with the 'Port' algorithm.
circan(data=expression_example, s2c=metadata_example, mode="port")
}
\keyword{CircaN}
\keyword{circadian}
\keyword{regression}
